Improve solutions

+++ areal.cpp
@@ -5,8 +5,10 @@
 int main() {
     long long a; 
     std::cin >> a;
-    long double result = (long double)std::sqrt(a) * (long double)4;
-    std::cout << std::setiosflags(std::ios::fixed) <<
-        std::setprecision(8) << result << std::endl;
+    long double result = (long double)std::sqrt(a) * 4;
+    std::cout << std::setiosflags(std::ios::fixed) 
+        << std::setprecision(8) 
+        << result 
+        << std::endl;
     return 0;
 }
